fix: AST improvements (#927-#931)
- Extract AST MCP handlers to ast_tools.rs (mod.rs: 2680->2195 lines) (#928)
- Wrap AST MCP handlers in spawn_blocking for HTTP transport (#927)
- Add PageRank convergence check (was fixed 20 iterations) (#927)
- Add verbose! output to all 11 CLI ast commands (#929)
- Add AST section to quickstart guide (#930)
- Report unsupported language name instead of 'No symbols found' (#931)

Closes #927
Closes #928
Closes #929
Closes #930
Closes #931
